JMK_102: Fetching proposals list from job posting id.

+++ src/main/java/com/jobmatrix/jm_proposal/dto/ProposalSubmissionDTO.java
@@ -5,9 +5,6 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.*;
-
-import java.math.BigDecimal;
-import java.util.List;
 import java.util.UUID;
 
 @Data
@@ -19,7 +16,7 @@ import java.util.UUID;
 public class ProposalSubmissionDTO {
 
     @NotNull(message = "Job Posting ID is required")
-    private int jobPostingId;
+    private Long jobPostingId;
 
     @NotNull(message = "Freelancer ID is required")
     private UUID freelancerId;
